RSE-905: Setup Custom fields for Prospect Case category using logic in Civicase. Rename upgraders
