+++ CRM/Prospect/Upgrader/Steps/Step1003.php
@@ -1,25 +1,31 @@
 <?php
 
-use CRM_Prospect_Setup_AddProspectCategoryCgExtendsValue as AddProspectCategoryCgExtendsValue;
-use CRM_Prospect_Setup_MoveCustomFieldsToProspecting as MoveCustomFieldsToProspecting;
+use CRM_Prospect_Setup_CreateProspectOwnerRelationship as CreateProspectOwnerRelationship;
+use CRM_Prospect_Setup_CreateProspectWorkflowCaseStatuses as CreateProspectWorkflowCaseStatuses;
+use CRM_Prospect_Setup_CreateProspectWorkflowCaseType as CreateProspectWorkflowCaseType;
 
 /**
- * Moves the prospect custom fields to the prospect category.
+ * Creates the Default Prospect Workflow case type.
+ *
+ * The Prospect owner relationship, the default prospect workflow statuses are
+ * also created.
+ *
+ * @return bool
+ *   Return value in boolean.
  */
 class CRM_Prospect_Upgrader_Steps_Step1003 {
 
   /**
-   * Moves the prospect custom fields to the prospect category.
-   *
-   * Adds the prospecting category as an extendable entity.
+   * Applies the setup functions.
    *
    * @return bool
-   *   Boolean value.
+   *   Return value.
    */
   public function apply() {
     $steps = [
-      new AddProspectCategoryCgExtendsValue(),
-      new MoveCustomFieldsToProspecting(),
+      new CreateProspectWorkflowCaseStatuses(),
+      new CreateProspectOwnerRelationship(),
+      new CreateProspectWorkflowCaseType(),
     ];
 
     foreach ($steps as $step) {
