+++ CRM/Prospect/Upgrader/Steps/Step1004.php
@@ -1,31 +1,25 @@
 <?php
 
-use CRM_Prospect_Setup_CreateProspectOwnerRelationship as CreateProspectOwnerRelationship;
-use CRM_Prospect_Setup_CreateProspectWorkflowCaseStatuses as CreateProspectWorkflowCaseStatuses;
-use CRM_Prospect_Setup_CreateProspectWorkflowCaseType as CreateProspectWorkflowCaseType;
+use CRM_Prospect_Setup_ProcessProspectCategoryForCustomGroupSupport as ProcessProspectCategoryForCustomGroupSupport;
+use CRM_Prospect_Setup_MoveCustomFieldsToWorkFlowCaseType as MoveCustomFieldsToWorkFlowCaseType;
 
 /**
- * Creates the Default Prospect Workflow case type.
- *
- * The Prospect owner relationship, the default prospect workflow statuses are
- * also created.
- *
- * @return bool
- *   Return value in boolean.
+ * Moves the prospect custom fields to the prospect category.
  */
 class CRM_Prospect_Upgrader_Steps_Step1004 {
 
   /**
-   * Applies the setup functions.
+   * Moves the prospect custom fields to the workflow case type.
+   *
+   * Adds the prospecting category as an extendable entity.
    *
    * @return bool
-   *   Return value.
+   *   Boolean value.
    */
   public function apply() {
     $steps = [
-      new CreateProspectWorkflowCaseStatuses(),
-      new CreateProspectOwnerRelationship(),
-      new CreateProspectWorkflowCaseType(),
+      new ProcessProspectCategoryForCustomGroupSupport(),
+      new MoveCustomFieldsToWorkFlowCaseType(),
     ];
 
     foreach ($steps as $step) {
